requirements.txt
fastapi==0.110.0
httpx==0.27.0
faiss-cpu==1.8.0
asyncio==3.4.3
tree-sitter==0.20.4
pydantic==2.7.1
uvicorn==0.27.0
python-dotenv==1.0.0